Refactor form field
